requirements.txt
google-genai==1.59.0
openai==2.15.0
python-dotenv
fastapi
uvicorn
pydantic
aiohttp
litellm